fix Ergo E2E test issue
