+++ app/api/ergo/lib/transactions/utils.js
@@ -108,6 +108,27 @@ export function asAddressedUtxo(
   });
 }
 
+export function toErgoBoxJSON(
+  utxos: Array<RemoteUnspentOutput>
+): Array<ErgoBoxJson> {
+  return utxos.map(utxo => {
+      return {
+        boxId: utxo.boxId,
+        value: Number.parseInt(utxo.amount, 10),
+        ergoTree: utxo.ergoTree,
+        assets: (utxo.assets ?? []).map(asset => ({
+          amount: asset.amount,
+          tokenId: asset.tokenId,
+        })),
+        creationHeight: utxo.creationHeight,
+        additionalRegisters: utxo.additionalRegisters || Object.freeze({}),
+        transactionId: utxo.tx_hash,
+        index: utxo.tx_index,
+      };
+    })
+}
+
+
 export function replaceMockBoxId(utxo: RemoteUnspentOutput): RemoteUnspentOutput {
   const tokens = new RustModule.SigmaRust.Tokens();
   for (const token of (utxo.assets ?? [])) {
