+++ package.json
@@ -108,7 +108,7 @@
     "crx": "5.0.1",
     "css-loader": "4.3.0",
     "cucumber": "git+https://github.com/SebastienGllmt/cucumber-js.git",
-    "ergo-lib-wasm-nodejs": "0.4.2",
+    "ergo-lib-wasm-nodejs": "0.4.4",
     "eslint": "7.12.1",
     "eslint-config-airbnb": "18.2.0",
     "eslint-config-prettier": "6.15.0",
@@ -173,7 +173,7 @@
     "classnames": "2.2.6",
     "crypto-random-string": "3.3.0",
     "dropbox": "4.0.30",
-    "ergo-lib-wasm-browser": "0.4.2",
+    "ergo-lib-wasm-browser": "0.4.4",
     "es6-error": "4.1.1",
     "file-saver": "1.3.8",
     "jdenticon": "3.0.1",
